feat: added before hooks that consumers can use for authorization until I have access rules working

+++ packages/generator/src/nexusTypes/mutations/deleteMany.ts
@@ -3,10 +3,12 @@ import { mutationField, nonNull } from 'nexus'
 import { isEmpty } from 'lodash'
 
 import { getNexusOperationArgs } from '../getNexusArgs'
+import { ModelDeleteConfiguration } from '../../_types/modelDeleteConfiguration'
 
 export const deleteMany = (
   modelName: string,
   mutationOutputTypes: DMMF.OutputType,
+  deleteConfig: ModelDeleteConfiguration,
   inputsWithNoFields:string[]
 ) => {
   const mutationName = `deleteMany${modelName}`
@@ -19,7 +21,15 @@ export const deleteMany = (
   return mutationField(mutationName, {
     type: nonNull('BatchPayload'),
     args,
-    resolve: async (_parent, { where }, { prisma }) => {
+    resolve: async (parent, args, ctx, info) => {
+      const { where } = args
+      const { prisma } = ctx
+
+      if (deleteConfig.beforeDeleteMany) {
+        const canDelete = deleteConfig.beforeDeleteMany(parent, args, ctx, info)
+        if (!canDelete) { throw new Error('Unauthorized') }
+      }
+
       return prisma[modelName].deleteMany({ where } as any)
     }
   })
